+++ Core/Src/matrix_driver.c
@@ -256,7 +256,7 @@ void drawInterpFrame(float factor)
 }
 ///////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-void setup_TIM8(uint8_t psc, uint8_t arr)
+void setup_TIM8(uint16_t psc, uint16_t arr)
 {
 	// Turn on the clock for timer 8
 	RCC  -> APB2ENR |= RCC_APB2ENR_TIM8EN; // TIM8 clock
